requirements.txt
daphne==4.1.2
Django==5.0.6
django-nested-admin==4.0.2
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.1
whitenoise==6.6.0
